anders comments addressed except for scheme
