+++ platform-operator/controllers/verrazzano/validator/ComponentValidatorImpl.go
@@ -5,7 +5,7 @@ package validator
 
 import (
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
+	//"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/transform"
 )
@@ -13,7 +13,8 @@ import (
 type ComponentValidatorImpl struct{}
 
 var _ v1alpha1.ComponentValidator = ComponentValidatorImpl{}
-var _ v1beta1.ComponentValidator = ComponentValidatorImpl{}
+
+//var _ v1beta1.ComponentValidator = ComponentValidatorImpl{}
 
 func (c ComponentValidatorImpl) ValidateInstall(vz *v1alpha1.Verrazzano) []error {
 	var errs []error
@@ -33,23 +34,23 @@ func (c ComponentValidatorImpl) ValidateInstall(vz *v1alpha1.Verrazzano) []error
 	return errs
 }
 
-func (c ComponentValidatorImpl) ValidateInstallV1Beta1(vz *v1beta1.Verrazzano) []error {
-	var errs []error
-
-	effectiveCR, err := transform.GetEffectiveCR(vz)
-	if err != nil {
-		errs = append(errs, err)
-		return errs
-	}
-
-	for _, comp := range registry.GetComponents() {
-		if err := comp.ValidateInstallV1Beta1(effectiveCR); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	return errs
-}
+//func (c ComponentValidatorImpl) ValidateInstallV1Beta1(vz *v1beta1.Verrazzano) []error {
+//	var errs []error
+//
+//	effectiveCR, err := transform.GetEffectiveCR(vz)
+//	if err != nil {
+//		errs = append(errs, err)
+//		return errs
+//	}
+//
+//	for _, comp := range registry.GetComponents() {
+//		if err := comp.ValidateInstallV1Beta1(effectiveCR); err != nil {
+//			errs = append(errs, err)
+//		}
+//	}
+//
+//	return errs
+//}
 
 func (c ComponentValidatorImpl) ValidateUpdate(old *v1alpha1.Verrazzano, new *v1alpha1.Verrazzano) []error {
 	var errs []error
@@ -73,24 +74,24 @@ func (c ComponentValidatorImpl) ValidateUpdate(old *v1alpha1.Verrazzano, new *v1
 	return errs
 }
 
-func (c ComponentValidatorImpl) ValidateUpdateV1Beta1(old *v1beta1.Verrazzano, new *v1beta1.Verrazzano) []error {
-	var errs []error
-
-	effectiveNew, err := transform.GetEffectiveCR(new)
-	if err != nil {
-		errs = append(errs, err)
-		return errs
-	}
-	effectiveOld, err := transform.GetEffectiveCR(old)
-	if err != nil {
-		errs = append(errs, err)
-		return errs
-	}
-
-	for _, comp := range registry.GetComponents() {
-		if err := comp.ValidateUpdateV1Beta1(effectiveOld, effectiveNew); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	return errs
-}
+//func (c ComponentValidatorImpl) ValidateUpdateV1Beta1(old *v1beta1.Verrazzano, new *v1beta1.Verrazzano) []error {
+//	var errs []error
+//
+//	effectiveNew, err := transform.GetEffectiveCR(new)
+//	if err != nil {
+//		errs = append(errs, err)
+//		return errs
+//	}
+//	effectiveOld, err := transform.GetEffectiveCR(old)
+//	if err != nil {
+//		errs = append(errs, err)
+//		return errs
+//	}
+//
+//	for _, comp := range registry.GetComponents() {
+//		if err := comp.ValidateUpdateV1Beta1(effectiveOld, effectiveNew); err != nil {
+//			errs = append(errs, err)
+//		}
+//	}
+//	return errs
+//}
